requirements.txt

streamlit
numpy
pandas
scikit-learn
matplotlib
seaborn
joblib
xgboost
imbalanced-learn
plotly
google-generativeai